Initial Commit.
This is the first commit of my own lisp, it follows a tutorial found
here: https://www.buildyourownlisp.com/chapter6_parsing

mpc.h and mpc.c are borrowed from here:
https://github.com/orangeduck/mpc

The license will be included in the next commit.

We'll se if I in the future will make this programming language to
something more or if I'll just leav it be.

In either case, if you're reading this thanks for your interest.